refactor to create a transformStyle function that only creates a MagicString when needed
